requirements.txt
dotenv
transformers[torch]
sentence-transformers
pillow
sqlite-vec